[#61] refactor: 이벤트 시작알람, 비밀번호 변경알람 기능 AlertService 클래스로 이전
