+++ src/main/java/com/festa/controller/MemberController.java
@@ -9,6 +9,7 @@ import com.festa.dto.MemberDTO;
 import com.festa.model.AlertResponse;
 import com.festa.model.MemberLogin;
 import com.festa.model.MemberInfo;
+import com.festa.service.AlertService;
 import com.festa.service.MemberService;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.log4j.Log4j2;
@@ -52,6 +53,7 @@ import static com.festa.common.ResponseEntityConstants.RESPONSE_ENTITY_OK;
 public class MemberController {
 
     private final MemberService memberService;
+    private final AlertService alertService;
     private final LoginService loginService;
     private final FirebaseTokenManager firebaseTokenManager;
 
@@ -133,8 +135,8 @@ public class MemberController {
 
         firebaseTokenManager.makeAccessToken(memberLogin.getUserNo());
 
-        List<AlertResponse> loginResponses = memberService.sendEventStartNotice(memberLogin.getUserNo(), LocalDate.now());
-        AlertResponse isUserNeedToChangePw = memberService.getChangePwDateDiff(memberLogin.getUserNo());
+        List<AlertResponse> loginResponses = alertService.sendEventStartNotice(memberLogin.getUserNo(), LocalDate.now());
+        AlertResponse isUserNeedToChangePw = alertService.getChangePwDateDiff(memberLogin.getUserNo());
 
         loginResponses.add(isUserNeedToChangePw);
