+++ src/test/java/com/festa/MemberControllerTests.java
@@ -7,6 +7,7 @@ import com.festa.controller.MemberController;
 import com.festa.dto.MemberDTO;
 import com.festa.model.MemberInfo;
 import com.festa.model.MemberLogin;
+import com.festa.service.AlertService;
 import com.festa.service.MemberService;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
@@ -43,6 +44,9 @@ class MemberControllerTests {
 
     @MockBean
     private FirebaseTokenManager firebaseTokenManager;
+
+    @MockBean
+    private AlertService alertService;
 
     private MockHttpSession mockHttpSession;
